ref: Move formatFloat out of formatters.tsx
