+++ static/app/utils/number/formatFloat.spec.tsx
@@ -1,0 +1,12 @@
+import {formatFloat} from 'sentry/utils/number/formatFloat';
+
+describe('formatFloat()', function () {
+  it('should format decimals', function () {
+    expect(formatFloat(0, 0)).toBe(0);
+    expect(formatFloat(10.513434, 1)).toBe(10.5);
+    expect(formatFloat(10.513494, 3)).toBe(10.513);
+  });
+  it('should not round', function () {
+    expect(formatFloat(10.513494, 4)).toBe(10.5134);
+  });
+});
