+++ static/app/utils/number/formatFloat.tsx
@@ -1,0 +1,4 @@
+export function formatFloat(number: number, places: number) {
+  const multi = Math.pow(10, places);
+  return parseInt((number * multi).toString(), 10) / multi;
+}
